==> tweet_classification/01_Twint_scrape.ipynb <==
Documentation: https://pypi.org/project/twint/

import pandas as pd
import numpy as np
import csv
from collections import Counter

# Read in 100 users per side
influential_csv = pd.read_csv('influential_tweeters.csv')

influential_csv.head(50)

influential_csv.shape

print(len(np.unique(influential_csv['denying'])))
print(len(np.unique(influential_csv['affirming'])))

print(len(influential_csv['denying']))
print(len(influential_csv['affirming']))

[x for x in Counter(influential_csv['denying']) if Counter(influential_csv['denying'])[x] > 1]

[x for x in Counter(influential_csv['affirming']) if Counter(influential_csv['affirming'])[x] > 1]

influential_csv['denying'].drop_duplicates(keep='first',inplace=True)
influential_csv['affirming'].drop_duplicates(keep='first',inplace=True)

influential_csv.shape

Scrape these users' tweets using Twint:

import twint
import nest_asyncio
import os
nest_asyncio.apply()

SEARCH_TERMS = ["sea","ice","melt","co2","methane","greenhouse","carbon",
               "emission","anthropogenic","temperature","warming","climate","coal","IPCC","environment",
               "environmental","cap and trade",'energy','planet','renewable','fossil fuel','environmentalist']
# TO DO: expand search_terms w/ some of Zach's terms

SCRAPE_DIR = '/per_user_search_term_scrapes'

def create_twint_config(user,end_date,search_term):
    c = twint.Config()
    c.Show_hashtags = True
    c.Count = True
    c.Stats = True
    c.Store_csv = True
    c.Retweets = False
    c.Custom["tweet"] = ["id", "date", "time", "tweet", "mentions", "replies_count", "retweets_count", 
                         "likes_count", "hashtags", "username"]
    c.Format = "ID {id} -||- Date {date} -||- Time {time} -||- Tweet {tweet} -||- Mentions {mentions} -||- \
    Repliescount {replies_count} \
    -||- Retweetscount {retweets_count} -||- Likescount {likes_count} -||- Hashtags {hashtags} -||- Username {username}"
    sep = " -||- "
    c.Username = user
    c.Until = end_date
    c.Search = search_term
    
    return c

# RESET

influential_csv.head(20)

curr_deny = influential_csv.iloc[92]['denying']
print('Current denying user:',curr_deny)
#c.Username = curr_affirm

## RESTART HERE

config = create_twint_config(
    curr_deny,None,None)

config.__dict__

if config.Search == None:
    start_index = 0
else:
    start_index = SEARCH_TERMS.index(config.Search)

start_index

for ix in range(21,len(SEARCH_TERMS)):
    search_term = SEARCH_TERMS[ix]
    #config.Until = None
    config.Search = search_term
    config.Output = os.getcwd()+SCRAPE_DIR+'/{}_search-term={}.csv'.format(config.Username,config.Search)
    print('Getting tweets with keyword: ' + config.Search)#search_term in SEARCH_TERMS:
    #c.Search = search_term
    #c.Pandas = True
    #c.Store_object = True
    #c.Output = os.getcwd()+SCRAPE_DIR+'/{}_search-term={}.csv'.format(c.Username,c.Search)
    twint.run.Search(config)
    

old_ix = ix

ix,search_term

latest_tweets = pd.read_csv(os.getcwd()+SCRAPE_DIR+'/{}_search-term={}.csv'.format(config.Username,search_term))
latest_date = latest_tweets['date'].iloc[-1]
config = create_twint_config(curr_deny,latest_date,config.Search)

Organize scraped Tweets from all searchterms in df; remove duplicates.

config.Username

user_tweets = []
for search_term in SEARCH_TERMS:
    try:
        tweets_per_search_term = pd.read_csv(os.getcwd()+SCRAPE_DIR+'/{}_search-term={}.csv'.format(config.Username,search_term))
        tweets_per_search_term['search_term'] = search_term
        user_tweets.append(tweets_per_search_term)
    except FileNotFoundError:
        pass

len(user_tweets) # each one is a df

user_tweets[0]

full_tweets = pd.concat(user_tweets,ignore_index=True)
full_tweets.shape

full_tweets['search_term'].value_counts()

#from collections import Counter
counted_ids = Counter(full_tweets['id'])
print(len([(id_,counted_ids[id_]) for id_ in counted_ids if counted_ids[id_] > 1]))

np.sum([counted_ids[id_] for id_ in counted_ids if counted_ids[id_] > 1])

full_tweets.drop_duplicates(subset ="id", keep = 'first', inplace = True)
full_tweets.shape

print(config.Username)

full_tweets.to_pickle('{}_unique_tweets.pkl'.format(config.Username))

full_tweets = pd.read_pickle('{}_unique_tweets.pkl'.format(config.Username))

full_tweets

import os
len(os.listdir(os.getcwd()+'/affirm_tweets/'))

os.listdir(os.getcwd()+'/deny_tweets/')

all_deny_tweets = []
for file in os.listdir(os.getcwd()+'/deny_tweets/'):
    tweets_per_user = pd.read_pickle(os.getcwd()+'/deny_tweets/'+file)
    all_deny_tweets.append(tweets_per_user)

deny_tweets = pd.concat(all_deny_tweets)
deny_tweets.shape

all_affirm_tweets = []
for file in os.listdir(os.getcwd()+'/affirm_tweets/'):
    tweets_per_user = pd.read_pickle(os.getcwd()+'/affirm_tweets/'+file)
    all_affirm_tweets.append(tweets_per_user)

affirm_tweets = pd.concat(all_affirm_tweets)
affirm_tweets.shape

deny_tweets.columns

%matplotlib inline

deny_tweets['username'].value_counts()[:10]

affirm_tweets['username'].value_counts()[:10]

engagement_df = pd.DataFrame.from_dict({'deny':
                                        {'retweets':
                                         {'mean':round(deny_tweets['retweets_count'].mean(),1),
                                         'min':round(deny_tweets['retweets_count'].min(),1),
                                         'max':round(deny_tweets['retweets_count'].max(),1)},
                                        'likes':
                                         {'mean':round(deny_tweets['likes_count'].mean(),1),
                                         'min':round(deny_tweets['likes_count'].min(),1),
                                         'max':round(deny_tweets['likes_count'].max(),1)}
                                        },
                                       'affirm':
                                        {
                                            'retweets':
                                            {'mean':round(affirm_tweets['retweets_count'].mean(),1),
                                         'min':round(affirm_tweets['retweets_count'].min(),1),
                                         'max':round(affirm_tweets['retweets_count'].max(),1)},
                                        'likes':
                                         {'mean':round(affirm_tweets['likes_count'].mean(),1),
                                         'min':round(affirm_tweets['likes_count'].min(),1),
                                         'max':round(affirm_tweets['likes_count'].max(),1)}
                                            
                                        }
                                       })
engagement_df

deny_tweets.loc[deny_tweets['retweets_count'] == deny_tweets['retweets_count'].max()]

deny_tweets.loc[deny_tweets['likes_count'] == deny_tweets['likes_count'].max()]

deny_tweets.loc[1089691985164800001]['tweet']

affirm_tweets.loc[affirm_tweets['retweets_count'] == affirm_tweets['retweets_count'].max()]

affirm_tweets = affirm_tweets.set_index('id')

affirm_tweets.loc[1151868495824642049]['tweet']

affirm_tweets.loc[affirm_tweets['likes_count'] == affirm_tweets['likes_count'].max()]

affirm_tweets.loc[1178002285751865344]['tweet']

deny_tweets.sort_values('retweets_count',ascending=False).head(1000)['username'].value_counts()

affirm_tweets.sort_values('retweets_count',ascending=False).head(1000)['username'].value_counts()

affirm_tweets.sort_values('likes_count',ascending=False).head(1000)['username'].value_counts()

print(deny_tweets['retweets_count'].mean())
print(deny_tweets['retweets_count'].max())
print(deny_tweets['retweets_count'].min())

print(deny_tweets['likes_count'].mean())
print(deny_tweets['likes_count'].max())
print(deny_tweets['likes_count'].min())

print(affirm_tweets['retweets_count'].mean())
print(affirm_tweets['retweets_count'].max())
print(affirm_tweets['retweets_count'].min())

print(affirm_tweets['likes_count'].mean())
print(affirm_tweets['likes_count'].max())
print(affirm_tweets['likes_count'].min())

deny_tweets.plot.hist(by='retweets_count',bins=10)
