==> tweet_classification/3.5_Log_odds-ethan.ipynb <==
import pickle
import numpy as np
import pandas as pd
from collections import Counter

# Functions for importing & cleaning relevant tweets
def lower(s):
    return s.lower()

def tweet_imports(filename):
    imp = pd.read_pickle(filename)
    imp = imp.drop_duplicates()
    imp['tweet_clean'] = imp['tweet'].str.replace('http\S+|www.\S+|pic.twitter.com\S+', '', case=False)
    imp['tweet_clean'] =imp['tweet_clean'].replace('[^A-Za-z0-9 ]+','',regex=True)
    imp['tweet_clean'] = imp['tweet_clean'].apply(lower)#map(lambda x: x.lower(), imp['tweet_clean'])
    imp['date'] = pd.to_datetime(imp['date'])
    return imp

def log_odds(l1,l2):
    counts_l1 = Counter(l1)
    counts_l2 = Counter(l2)
    counts_l1.update({t: 0.5 for t in counts_l2 if t not in counts_l1})
    counts_l2.update({t: 0.5 for t in counts_l1 if t not in counts_l2})
    sum_1 = sum(counts_l1.values())
    sum_2 = sum(counts_l2.values())
    freqs_1 = {t: counts_l1[t]*1./sum_1 for t in counts_l1}
    freqs_2 = {t: counts_l2[t]*1./sum_2 for t in counts_l2}
    odds_1 = {t: freqs_1[t]/(1-freqs_1[t]) for t in freqs_1}
    odds_2 = {t: freqs_2[t]/(1-freqs_2[t]) for t in freqs_2}
    odds_ratios = {t: odds_1[t]/odds_2[t] for t in odds_1}
    return {t: np.log(odds_ratios[t]) for t in odds_ratios}

from nltk.tokenize import MWETokenizer
tk = MWETokenizer()
tk.add_mwe(('climate','change'))
tk.add_mwe(('global','warming'))
tk.add_mwe(('one','world'))
tk.add_mwe(('new','jobs'))
tk.add_mwe(('carbon','tax'))
tk.add_mwe(('carbon','neutral'))

tk.tokenize('Global warming is exaggerated'.lower().split())

import nltk

from nltk.stem.porter import *
stemmer = PorterStemmer()

import re

liwc_dict = {}
with open('/Users/yiweiluo/Downloads/datasets/en_liwc.txt','r') as f:
    for line in f:
        split_line = line.strip().split(': ')
        cat = split_line[0]
        cat_words = split_line[1].split()
        cat_words = [w.split('*')[0] for w in cat_words]
        liwc_dict[cat] = cat_words

liwc_dict.keys()

pd.DataFrame.from_dict(liwc_dict,orient='index')

RELEVANT_LIWC_CATS = ['Inhib','Affect','Cause','Relig','Body','Bio','Posemo','Home','Work','Anger','Social',
                     'Family','Humans','Sad','Future','Past','Money','Negate','Health','Negemo']

def log_odds_liwc(odds_dict,liwc_cat):
    liwc_words = liwc_dict[liwc_cat]
    liwc_odds = {w: odds_dict[w] for w in liwc_words if w in odds_dict}
    return liwc_odds

# News data

news_df = pd.read_pickle('/Users/yiweiluo/Dropbox/research/QP2/code/Fox_and_friends/\
all_urls_meta_and_fulltext_df_2020.pkl')

news_df = news_df.reset_index()
climate_news = news_df[news_df['topic'] == 'cc']

climate_news.head()

climate_news[climate_news['stance'] == 'anti'].shape

Tokenize and pre-process fulltext from each side:

pro_news_tokens = [tk.tokenize(climate_news.iloc[row_ix]['fulltext'].lower().split()) 
                   for row_ix in range(len(climate_news.index))
                  if climate_news.iloc[row_ix]['fulltext'] is not None and 
                   len(climate_news.iloc[row_ix]['fulltext']) > 0 and climate_news.iloc[row_ix]['stance'] == 'pro']
anti_news_tokens = [tk.tokenize(climate_news.iloc[row_ix]['fulltext'].lower().split()) 
                    for row_ix in range(len(climate_news.index))
                  if climate_news.iloc[row_ix]['fulltext'] is not None and 
                    len(climate_news.iloc[row_ix]['fulltext']) > 0 and climate_news.iloc[row_ix]['stance'] == 'anti']

pro_news_tokens = [item for sublist in pro_news_tokens for item in sublist]
anti_news_tokens = [item for sublist in anti_news_tokens for item in sublist]

ethan_pro_news_tokens = pickle.load(open('pro_news_tokens.pkl','rb'))
print("PRO NEWS TOKENS:")
print(ethan_pro_news_tokens[:100])
ethan_anti_news_stems = pickle.load(open('anti_news_stems.pkl','rb'))
# anti_words = []
# for word in ethan_anti_news_stems:
#     if (not word.isalpha()) and (not word.isnumeric()):
#         anti_words.append(word)
# print anti_words[:500]
print("ANTI NEWS STEMS:")
print(ethan_anti_news_stems[:100])

Remove punctuation from tokens and get stems:

import string
string.punctuation
punctuations = string.punctuation + '“'

string.punctuation

import re

def cleanItems(items):
    cleanedItems = []
    for item in items:
        item = re.sub('[^a-zA-Z0-9 ]', '', item)
        if item:
            cleanedItems.append(item)
    return cleanedItems

print("CLEANING...")
ethan_cleaned_pro = cleanItems(ethan_pro_news_tokens)
ethan_cleaned_anti = cleanItems(ethan_anti_news_stems)

print("CLEANED PRO NEWS TOKENS:")
print(ethan_cleaned_pro[:500])

print("CLEANED ANTI NEWS STEMS:")
print(ethan_cleaned_anti[:500])

pickle.dump(ethan_cleaned_pro, open('ethan_pro_news_tokens.pkl','wb'))
pickle.dump(ethan_cleaned_anti, open('ethan_anti_news_stems.pkl','wb'))
    

pro_news_tokens = [item.translate(str.maketrans('', '', punctuations))
                   for item in pro_news_tokens]
anti_news_tokens = [item.translate(str.maketrans('', '', punctuations))
                   for item in anti_news_tokens]

from spacy.lemmatizer import Lemmatizer
#from spacy.lookups import Lookups
#lookups = Lookups()
#lookups.add_table("lemma_rules", {"noun": [["s", ""]]})
lemmatizer = Lemmatizer()
lemmas = lemmatizer("ducks", "NOUN")

import nltk
from nltk.stem import WordNetLemmatizer
lemmatizer = WordNetLemmatizer()

lemmatizer.lemmatize('global_warming')

lemmatizer.lemmatize('ducks')

# pro_news_lemmas = [lemmatizer.lemmatize(item) for item in pro_news_tokens]
# anti_news_lemmas = [lemmatizer.lemmatize(item) for item in anti_news_tokens]
pro_news_stems = [stemmer.stem(item) for item in pro_news_tokens]
anti_news_stems = [stemmer.stem(item) for item in anti_news_tokens]

anti_news_lemmas[:50]

print(len(pro_news_lemmas))
print(len(anti_news_lemmas))

# pickle.dump(pro_news_lemmas,open('pro_news_lemmas.pkl','wb'))
# pickle.dump(anti_news_lemmas,open('anti_news_lemmas.pkl','wb'))
pickle.dump(pro_news_stems,open('pro_news_stems.pkl','wb'))
pickle.dump(anti_news_stems,open('anti_news_stems.pkl','wb'))

#pro_anti_news_odds = log_odds(pro_news_lemmas,anti_news_lemmas)
pro_anti_news_odds_stems = log_odds(pro_news_stems,anti_news_stems)

#sorted_pro_anti_news_odds = sorted(pro_anti_news_odds.items(), key=lambda x:x[1])
sorted_pro_anti_news_odds_stems = sorted(pro_anti_news_odds_stems.items(), key=lambda x:x[1])

sorted_pro_anti_news_odds[-100:]

#dict_odds = dict(pro_anti_news_odds)
dict_odds_stems = dict(pro_anti_news_odds_stems)

dict_odds['cause']

list(dict_odds.keys())[:5]

dict_odds['matter']

## Of specific LIWC categories

for cat in RELEVANT_LIWC_CATS:
    odds = log_odds_liwc(dict_odds_stems,cat)
    pos = [w for w in odds if odds[w] > 0]
    neg = [w for w in odds if odds[w] < 0]
    print(cat,'Num pos:'+str(len(pos)),'Num neg:'+str(len(neg)))
    print('\n')

sorted(log_odds_liwc(dict_odds_stems,'Inhib').items(),key=lambda x:x[1],reverse=True)

sorted(log_odds_liwc(dict_odds_stems,'Family').items(),key=lambda x:x[1],reverse=True)

sorted(log_odds_liwc(dict_odds_stems,'Health').items(),key=lambda x:x[1],reverse=True)

# Twitter data

pro_sentences_fuzzy_matched = pickle.load(open('fuzzy_matched_pro_sents.pkl','rb'))
anti_sentences_fuzzy_matched = pickle.load(open('fuzzy_matched_anti_sents.pkl','rb'))

anti_sentences_fuzzy_matched[:5]

cleaned_affirm_tweets

cleaned_affirm_tweets = tweet_imports('all_affirm_tweets.pkl')
cleaned_deny_tweets = tweet_imports('all_deny_tweets.pkl')

pro_sentences = [tk.tokenize(tweet.lower().split()) for tweet in cleaned_affirm_tweets['tweet_clean']]
anti_sentences = [tk.tokenize(tweet.lower().split()) for tweet in cleaned_deny_tweets['tweet_clean']]

# TO DO: fuzzy matching after MWE tokenization

fuzzy_pro_toks = [item for sublist in pro_sentences_fuzzy_matched for item in sublist]
fuzzy_anti_toks = [item for sublist in anti_sentences_fuzzy_matched for item in sublist]
#pro_toks = [item for sublist in pro_sentences for item in sublist if len(item) < 15]
#anti_toks = [item for sublist in anti_sentences for item in sublist if len(item) < 15]

pro_sentences[:5]

#pro_anti_log_odds_ratios = log_odds(pro_toks,anti_toks)
fuzzy_pro_anti_log_odds_ratios = log_odds(fuzzy_pro_toks,fuzzy_anti_toks)

#sorted_pro_anti_log_odds_ratios = sorted(pro_anti_log_odds_ratios.items(),key=lambda x:x[1],reverse=True)
fuzzy_sorted_pro_anti_log_odds_ratios = sorted(fuzzy_pro_anti_log_odds_ratios.items(),key=lambda x:x[1],reverse=True)

sorted_pro_anti_log_odds_ratios[:50]

fuzzy_sorted_pro_anti_log_odds_ratios[:50]

sorted_pro_anti_log_odds_ratios[-50:]

fuzzy_sorted_pro_anti_log_odds_ratios[-50:]
