==> tweet_classification/5_Deconfounded_Lexicon_Induction.ipynb <==
import pickle
import numpy as np
import pandas as pd
from keras import Sequential, backend, regularizers
from keras.layers import Embedding, LSTM, Dense, Dropout, Flatten, Conv1D, MaxPooling1D
from keras.preprocessing import sequence
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
import keras.backend
import os

pd.options.display.max_rows = 999
pd.options.display.max_columns = 999

# Load tweet data

# Functions for importing & cleaning relevant tweets
def lower(s):
    return s.lower()

def tweet_imports(filename):
    imp = pd.read_pickle(filename)
    imp = imp.drop_duplicates()
    imp['tweet_clean'] = imp['tweet'].str.replace('http\S+|www.\S+|pic.twitter.com\S+', '', case=False)
    imp['tweet_clean'] =imp['tweet_clean'].replace('[^A-Za-z0-9 ]+','',regex=True)
    imp['tweet_clean'] = imp['tweet_clean'].apply(lower)#map(lambda x: x.lower(), imp['tweet_clean'])
    imp['date'] = pd.to_datetime(imp['date'])
    return imp

cleaned_affirm_tweets = tweet_imports('all_affirm_tweets.pkl')
cleaned_deny_tweets = tweet_imports('all_deny_tweets.pkl')

cleaned_affirm_tweets.head()

len(cleaned_affirm_tweets)

len(cleaned_deny_tweets)

import csv
# with open('generated.tsv','w') as f:
#     csvwriter = csv.writer(f,delimiter='\t')
#     for ix in range(100):#len(cleaned_deny_tweets)):
#         df_row = cleaned_deny_tweets.iloc[ix]
#         csvwriter.writerow([df_row['tweet_clean'],float(df_row['likes_count']),
#                                   float(df_row['retweets_count']),float(df_row['replies_count']),
#                             df_row['username']])
#     for ix in range(100):#len(cleaned_deny_tweets)):
#         df_row = cleaned_deny_tweets.iloc[-ix]
#         csvwriter.writerow([df_row['tweet_clean'],float(df_row['likes_count']),
#                                   float(df_row['retweets_count']),float(df_row['replies_count']),
#                             df_row['username']])

with open('generated_affirm.tsv','w') as f:
    csvwriter = csv.writer(f,delimiter='\t')
    for ix in range(100):#len(cleaned_deny_tweets)):
        df_row = cleaned_affirm_tweets.iloc[ix]
        csvwriter.writerow([df_row['tweet_clean'],float(df_row['likes_count']),
                                  float(df_row['retweets_count']),float(df_row['replies_count']),
                            df_row['username']])
    for ix in range(100):#len(cleaned_deny_tweets)):
        df_row = cleaned_affirm_tweets.iloc[-ix]
        csvwriter.writerow([df_row['tweet_clean'],float(df_row['likes_count']),
                                  float(df_row['retweets_count']),float(df_row['replies_count']),
                            df_row['username']])
